+++ src/sprout/Helpers/CustomHeadTags.php
@@ -97,47 +97,6 @@ class CustomHeadTags
 
 
     /**
-     * Get the list of tags set for the current page
-     *
-     * @param int $page_id The page ID to get the tags for
-     *
-     * @return array
-     */
-    public static function getHomepageTags(int $homepage_id): array
-    {
-        $q = "SELECT * FROM ~homepage_custom_tags WHERE homepage_id = ?";
-        $tags = Pdb::query($q, [$homepage_id], 'arr');
-
-        foreach ($tags as &$tag) {
-            $tag['attr_values'] = json_decode($tag['attr_values'], true);
-        }
-        unset($tag);
-
-        return $tags;
-    }
-
-
-    /**
-     * Render the form element for the custom meta tags on the homepage
-     *
-     * @param int $homepage_id The homepage ID to get the tags for
-     *
-     * @return void
-     */
-    public static function renderTagsFormElementHome(int $homepage_id): void
-    {
-        $available_tags = static::getAvailableTagList();
-        $current_tags = static::getHomepageTags($homepage_id);
-
-        $view = new PhpView('sprout/views/admin/custom_head_tag_edit');
-        $view->available_tags = $available_tags;
-        $view->current_tags = $current_tags;
-
-        echo $view->render();
-    }
-
-
-    /**
      * Render the form element for the custom meta tags
      *
      * @param int $page_id The page ID to get the tags for
@@ -178,33 +137,13 @@ class CustomHeadTags
 
 
     /**
-     * Save the custom meta tags for the given page
-     *
-     * @param int $page_id The page ID to save the tags for
-     * @param array $tags The tags to save
-     *
-     * @return void
-     */
-    public static function saveHomepageTags(int $page_id, array $tags): void
-    {
-        Pdb::delete('homepage_custom_tags', ['homepage_id' => $page_id]);
-
-        $data = static::buildTagsData($tags);
-        foreach ($data as $insert) {
-            $insert['homepage_id'] = $page_id;
-            Pdb::insert('homepage_custom_tags', $insert);
-        }
-    }
-
-
-    /**
      * Build the data array for saving the tags, excluding the (home)page ID
      *
      * @param array $tags The tags to save
      *
      * @return array
      */
-    private static function buildTagsData(array $tags): array
+    protected static function buildTagsData(array $tags): array
     {
         $available_tags = static::getAvailableTagList();
 
@@ -238,18 +177,6 @@ class CustomHeadTags
         if (!$node) return;
 
         $tags = static::getPageTags($node['id']);
-        static::addTagNeeds($tags);
-    }
-
-
-    /**
-     * Render the custom meta tags for the current homepage
-     *
-     * @return void
-     */
-    public static function addHeadTagsHome(int $homepage_id): void
-    {
-        $tags = static::getHomepageTags($homepage_id);
         static::addTagNeeds($tags);
     }
 
@@ -330,7 +257,7 @@ class CustomHeadTags
      *
      * @return void
      */
-    private static function addTagNeeds(array $tags): void
+    protected static function addTagNeeds(array $tags): void
     {
         foreach ($tags as $tag) {
             // Canonical handled in Page Controller
